angelsrefining: Create a new default map generation setting
This was intended to replace the "Angel's enemies slider", but currently
this is not feasible.

In space age, we actually can't do what the slider did and fixup the
enemy spawn rate to 0, as you need spawners to progress (capture a
spawner). The alternative would be peaceful mode or no enemies mode.
However neither of these can be enabled by default, they can only be
changed in (non-default) presets. That leaves this approach impossible.